jwt_drf_app/serializers.py
from jwt_drf_app.models import Post
from rest_framework import serializers
from django.contrib.auth.models import User
from django.contrib.auth.hashers import make_password


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ("username", "password")
        extra_kwargs = {"password": {"write_only": True}}

    def create(self, validated_data):
        user = User(username=validated_data["username"],
                    password=make_password(validated_data["password"]))
        user.save()
        return user


class PostSerializer(serializers.ModelSerializer):
    class Meta:
        model = Post
        fields = ("id", "author", "created", "title", "body", "likes_count")

    author = serializers.ReadOnlyField(source="author.username")
