requirements.txt
Django==2.2.2
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
psycopg2==2.8.3
requests==2.22.0
